fix: Improve token metadata update process
