+++ apps/indexer/lib/indexer/fetcher/token_updater.ex
@@ -52,7 +52,7 @@ defmodule Indexer.Fetcher.TokenUpdater do
       |> Duration.to_minutes()
       |> trunc()
 
-    {:ok, tokens} = Chain.stream_cataloged_tokens(initial, reducer, interval_in_minutes, true)
+    {:ok, tokens} = Token.stream_cataloged_tokens(initial, reducer, interval_in_minutes, true)
 
     tokens
   end
@@ -88,6 +88,6 @@ defmodule Indexer.Fetcher.TokenUpdater do
   end
 
   def update_metadata(%Token{} = token, metadata) do
-    Chain.update_token(token, metadata)
+    Token.update(token, metadata)
   end
 end
